update userList api to userInfo api and solved memberId problem

+++ backend/src/main/java/com/greenhair/template/service/JwtService.java
@@ -5,7 +5,7 @@ import java.util.Map;
 public interface JwtService {
 	<T> String create(String key, T data, String subject);
 	Map<String, Object> get(String key);
-	int getMemberId();
+	long getMemberId();
 	boolean isUsable(String jwt);
 	
 }
